Move `Blocks` and `Block` to dedicated module
