+++ crates/fj-kernel/src/stores/mod.rs
@@ -1,5 +1,6 @@
 //! Append-only object storage
 
+mod blocks;
 mod store;
 
 use crate::objects::GlobalCurve;
